File cleanup for ease of future reference
I learned a lot from this app: Spent a lot of time studying migrations, referencial integrity, Devise internals,
Rails views & partials, Rspec (but just routing xD), human readable urls (I wasted so much time on this),
semantic CSS and the tradeoffs between modern best practices (flexbox) and browser support.

In the beginning, this app was supposed to be so much more than this, but I was never able to identify exactly
what that was, and drive and ambition gave way to doubt and minutiae. Still, I'm grateful for everything I
learned from this experience.